feat(postcss): Prevent prefixing of :root selector
This commit addresses an issue where the `postcss-prefix-selector` plugin was incorrectly
applying prefixes to the `:root` selector.

By adding `:root` to the `exclude` option within the `postcssPrefixSelector` configuration
for both `cssanimation.css` and `cssanimation-utility.css`, we ensure that this global
selector remains un-prefixed, preserving its intended functionality.

This change prevents unexpected styling issues and improves the reliability of the
CSS output.